updatable receiver subgraph
