+++ packages/subgraph/src/mappings/01/utils.ts
@@ -1,18 +1,11 @@
-import {
-  Address,
-  BigInt,
-  Bytes,
-  ByteArray,
-  ipfs,
-  json,
-  log,
-} from '@graphprotocol/graph-ts';
+import { Address, BigInt, Bytes, ByteArray } from '@graphprotocol/graph-ts';
 
 import { Agreement, Invoice } from '../../types/schema';
 
 import { updateEscrowInfo } from './helpers/escrow';
 import { updateInstantInfo } from './helpers/instant';
 import { updateSplitEscrowInfo } from './helpers/split-escrow';
+import { updateUpdatableInfo } from './helpers/updatable-escrow';
 
 let zeroAddress = changetype<Address>(
   Address.fromHexString('0x0000000000000000000000000000000000000000'),
@@ -49,10 +42,12 @@ export class InvoiceObject {
   fulfilled: boolean;
   dao: Address;
   daoFee: BigInt;
+  providerReceiver: Address;
 
   constructor() {
     this.client = zeroAddress;
     this.provider = zeroAddress;
+    this.providerReceiver = zeroAddress;
     this.resolverType = 0;
     this.resolver = zeroAddress;
     this.resolutionRate = BigInt.fromI32(0);
@@ -100,6 +95,8 @@ export function updateInvoice(address: Address, invoice: Invoice): Invoice {
         invoice = updateEscrowInfo(address, invoice);
       } else if (type == 'split-escrow') {
         invoice = updateSplitEscrowInfo(address, invoice);
+      } else if (type == 'updatable') {
+        invoice = updateUpdatableInfo(address, invoice);
       } else {
         invoice = updateInstantInfo(address, invoice);
       }
